extract all hooks -> hooks folder
