+++ src/app/hooks/useCustomCourseSelection.ts
